requirements.txt
Pillow==10.3.0
protobuf==5.26.1
python-dotenv==1.0.1
streamlit==1.32.2
youtube_transcript_api==0.6.2
google
